improve AssetList
I realised I was saving the list of assets in the component state unnecessarily, I am instead using `props.assetList` directly and seperated the filters for `activeAssets` and `searchQuery`, which fixed a bug in which the data being console logged is filtered by the `searchQuery` value. This is past the 8 hour mark so I've created a PR instead of pushing to master.