desktop styling improvement
I noticed that the right side panel was taking up too much space in the UI on larger screens, I set the max-width to 284px to more accurately reflect the design. I have also removed the max-width of the search bar so it takes up the full width available on desktop.